jwt child deletion with authorisation

+++ src/solids_food_tracker_api/controller/children_controller.py
@@ -2,9 +2,10 @@ from flask import Blueprint, request, abort, jsonify
 from main import db
 from model.child import Child
 from model.food_tracker import FoodTracker
+from model.parent import Parent
 from schema.children_schema import child_schema, children_schema
 from schema.food_trackers_schema import food_tracker_schema, food_trackers_schema
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 
 child = Blueprint('child', __name__, url_prefix="/children")
@@ -12,16 +13,12 @@ child = Blueprint('child', __name__, url_prefix="/children")
 #child 
 
 @child.get("/")
-#Decorator to make sure the jwt is included in the request
-@jwt_required()
 def get_children():
     children = Child.query.all()
     return children_schema.dump(children)
 
 
 @child.get("/<int:id>")
-#Decorator to make sure the jwt is included in the request
-@jwt_required()
 def get_child(id):
     child = Child.query.get(id)
 
@@ -32,8 +29,6 @@ def get_child(id):
 
 
 @child.post("/")
-#Decorator to make sure the jwt is included in the request
-@jwt_required()
 def create_child():
 
     child_fields = child_schema.load(request.json)
@@ -47,8 +42,6 @@ def create_child():
 
 
 @child.put("/<int:id>")
-#Decorator to make sure the jwt is included in the request
-@jwt_required()
 def update_child(id):
 
     child_fields = child_schema.load(request.json)
@@ -70,17 +63,27 @@ def update_child(id):
 #Decorator to make sure the jwt is included in the request
 @jwt_required()
 def delete_child(id):
+    # delete child --> request send the child id in the url
+    # as the route is authenticated I have the parent id in the JWT token
+    # to confirm if the child is accessible by the parent, i simply find the child and check if the parentid is the same
 
+    #get the user id invoking get_jwt_identity
+    parent_id = int(get_jwt_identity())
+
+    #find the child
     child = Child.query.get(id)
 
     if not child:
         return abort(400, description= "No deletion. Child does not exist")
     
+    if child.parent_id != parent_id:
+        return abort(401, description= "No deletion. Parent does not have access to this child")
+
+    
     db.session.delete(child)
     db.session.commit()
     
     return {"message": "Child deleted!"}
-
 
 
 #food tracker
